// generated/environments.ts
// This file was auto-generated by Fern from our API Definition.

export const VoltariaEnvironment = {
    Sandbox: "https://api.sandbox.voltaria.io",
    Production: "https://api.voltaria.io",
} as const;

export type VoltariaEnvironment = typeof VoltariaEnvironment.Sandbox | typeof VoltariaEnvironment.Production;